+++ lib/nsrr/helpers/hash_helper.rb
@@ -1,0 +1,9 @@
+module Nsrr
+  module Helpers
+    class HashHelper
+      def self.symbolize_keys(hash)
+        hash.inject({}){|memo,(k,v)| memo[k.to_sym] = v; memo}
+      end
+    end
+  end
+end