+++ lib/nsrr/models/dataset.rb
@@ -1,6 +1,8 @@
 require 'colorize'
+require 'fileutils'
 
 require 'nsrr/helpers/constants'
+require 'nsrr/helpers/hash_helper'
 require 'nsrr/helpers/json_request'
 
 require 'nsrr/models/file'
@@ -23,9 +25,10 @@ module Nsrr
         @slug = json['slug']
         @name = json['name']
         @files = {}
+        @download_token = nil
       end
 
-      def files(path = nil) # TODO CHANGE TO nil
+      def files(path = nil)
         @files[path] ||= begin
           json = Nsrr::Helpers::JsonRequest.get("#{Nsrr::WEBSITE}/datasets/#{@slug}/json_manifest/#{path}")
           (json || []).collect{|file_json| Nsrr::Models::File.new(file_json)}
@@ -44,27 +47,75 @@ module Nsrr
       # depth:
       #   'recursive' => [default] Downloads files in selected path folder and all subfolders
       #   'shallow'   => Only downloads files in selected path folder
-      def download(path = nil, options = {})
-        method = options['method'] || 'md5'
-        depth = options['depth'] || 'recursive'
+      def download(path = nil, *args)
+        options = Nsrr::Helpers::HashHelper.symbolize_keys(args.first || {})
+        options[:method] ||= 'md5'
+        options[:depth] ||= 'recursive'
+        @folders_created = 0
+        @files_downloaded = 0
+        @downloaded_bytes = 0
+        @files_skipped = 0
+        @files_failed = 0
 
-        self.files(path).each do |file|
-          if file.is_file
-            file.download(method)
+        begin
+          if @download_token.to_s == ''
+            puts "     File Integrity Check Method: " + options[:method].to_s.colorize(:white)
+            puts "                           Depth: " + options[:depth].to_s.colorize(:white)
+            set_download_token()
+          end
+
+          @start_time = Time.now
+
+          download_helper(path, options)
+        rescue Interrupt, IRB::Abort
+          puts "\n   Interrupted".colorize(:red)
+        end
+
+        @downloaded_megabytes = @downloaded_bytes / (1024 * 1024)
+
+        puts "\nFinished in #{Time.now - @start_time} seconds."
+        puts "\n#{@folders_created} folder#{"s" if @folders_created != 1} created".colorize(:white) + ", " +
+             "#{@files_downloaded} file#{"s" if @files_downloaded != 1} downloaded".colorize(:green) + ", " +
+             "#{@downloaded_megabytes} MiB#{"s" if @downloaded_megabytes != 1} downloaded".colorize(:green) + ", " +
+             "#{@files_skipped} file#{"s" if @files_skipped != 1} skipped".colorize(:blue) + ", " +
+             "#{@files_failed} file#{"s" if @files_failed != 1} failed".colorize(:red) + "\n\n"
+        nil
+      end
+
+      def download_helper(path, options)
+        current_folder = ::File.join(self.slug.to_s, path.to_s)
+        create_folder(current_folder)
+
+        self.files(path).select{|f| f.is_file}.each do |file|
+          result = file.download(options[:method], current_folder, @download_token)
+          case result when 'fail'
+            @files_failed += 1
+          when 'skip'
+            @files_skipped += 1
           else
-            # puts "Folder Not Downloaded"
+            @files_downloaded += 1
+            @downloaded_bytes += result
           end
         end
 
-        if depth == 'recursive'
-          self.files(path).each do |file|
-            unless file.is_file
-              folder = [path, file.name].compact.join('/')
-              puts "      create".colorize( :white ) + " #{folder}"
-              self.download(folder, options)
-            end
+        if options[:depth] == 'recursive'
+          self.files(path).select{|f| !f.is_file}.each do |file|
+            folder = [path, file.name].compact.join('/')
+            self.download_helper(folder, options)
           end
         end
+      end
+
+      def create_folder(folder)
+        puts "      create".colorize( :white ) + " #{folder}"
+        FileUtils.mkdir_p folder
+        @folders_created += 1
+      end
+
+      def set_download_token
+        puts  "             Get your token here: " + "https://sleepdata.org/token".colorize(:blue)
+        print "Please enter your download token: "
+        @download_token = gets.chomp
       end
 
     end