+++ test/helpers/constants_test.rb
@@ -1,0 +1,16 @@
+require 'test_helper'
+require 'test_helpers/capture'
+
+require 'nsrr/helpers/constants'
+
+module HelpersTests
+  class ConstantsTest < Minitest::Test
+
+    include TestHelpers::Capture
+
+    def test_site_constant
+      assert_equal 'https://sleepdata.org', Nsrr::WEBSITE
+    end
+
+  end
+end